add semaphore implementation
implementation notes:

semaphore will release after a timeout, but does not
stop the timed out client from making additional changes -
it only keeps the semaphore from blocking up with locked
clients.

because this semaphore is designed for use with sidekiq
workers, it raises an exception if the client is unable to
obtain a lock. the api was built this way to force clients
to handle the fail case without blocking.